Determine Push Constants Per Frame to Set The Sample Window
GroupsPerFrame determines throughput at the moment.

It'll probably be beneficial to change this such that a thread group cranks
through the accumulation for one sample in tandem, since the GPU is
underutilized with the current scheme when GroupsPerFrame is low.

It really depends on how many samples must be processed at once though.